pos_customer_display: support ePOS display protocol
In addition to the USB LED LCD connected to IoTBox or pywebdriver, this
module now also the Epson ePOS protocol for displays: if you have an
Epson USB display (DM-D30 for example) connected to an Epson IP printer
that support ePOS (TM-m30 for example) and has ePOS print enabled in
it's web configuration interface, Odoo can connect directly to it. The
JS code of the POS will send the IP request to the ePOS printer that
will relay the information to the USB display.

Advantages of ePOS vs IoTBox/pywebdriver:
- supports accentuated caracters
- no need to install/setup IoTBox/pywebdriver

I added a dependency on pos_epson_printer, but it should not be a real
problem because it is a very small module and it won't bother you if you
don't use it.

Remove demo data because there are 2 options now for LCD on pos.config, so we let the user choose.

Use 'x' instead of '*' for multiple quantities.